+++ boilerplates/nuxt/components/Divider/DividerBase.vue
@@ -1,5 +1,5 @@
 <template>
-  <hr class="h-px bg-gray-200 w-80">
+  <hr class="w-full h-px bg-gray-200">
 </template>
 
 <script>