+++ boilerplates/nuxt/components/Spinner/SpinnerBase.vue
@@ -1,6 +1,14 @@
 <template>
   <div
-    class="relative h-6"
+    class="relative"
+    :class="{
+      'w-4 h-4': size === 'sm',
+      'w-6 h-6': size === 'base',
+      'w-10 h-10': size === 'lg',
+      'w-14 h-14': size === 'xl',
+      'w-24 h-24': size === '2xl',
+      'w-48 h-48': size === '3xl',
+    }"
     role="progressbar"
   >
     <div
@@ -10,23 +18,63 @@
       loading...
     </div>
     <div
-      class="absolute w-6 h-6 bg-transparent border-2 border-gray-200 rounded-full"
+      :class="{
+        'w-4 h-4 border-2': size === 'sm',
+        'w-6 h-6 border-2': size === 'base',
+        'w-10 h-10 border-2': size === 'lg',
+        'w-14 h-14 border-[3px]': size === 'xl',
+        'w-24 h-24 border-4': size === '2xl',
+        'w-48 h-48 border-8': size === '3xl',
+      }"
+      class="absolute bg-transparent border-gray-200 rounded-full"
       aria-hidden="true"
     />
     <div
-      class="absolute w-6 h-6 bg-transparent border-2 rounded-[50%] is-animation-stopped animate-spin-base border-t-primary-600"
+      :class="{
+        'w-4 h-4 border-2': size === 'sm',
+        'w-6 h-6 border-2': size === 'base',
+        'w-10 h-10 border-2': size === 'lg',
+        'w-14 h-14 border-[3px]': size === 'xl',
+        'w-24 h-24 border-4': size === '2xl',
+        'w-48 h-48 border-8': size === '3xl',
+      }"
+      class="absolute bg-transparent rounded-[50%] is-animation-stopped animate-spin-base border-t-primary-600"
       aria-hidden="true"
     />
     <div
-      class="absolute w-6 h-6 bg-transparent border-2 rounded-[50%] is-animation-stopped animate-spin-base animation-delay-1 border-t-primary-600 border-x-transparent border-b-transparent"
+      :class="{
+        'w-4 h-4 border-2': size === 'sm',
+        'w-6 h-6 border-2': size === 'base',
+        'w-10 h-10 border-2': size === 'lg',
+        'w-14 h-14 border-[3px]': size === 'xl',
+        'w-24 h-24 border-4': size === '2xl',
+        'w-48 h-48 border-8': size === '3xl',
+      }"
+      class="absolute bg-transparent rounded-[50%] is-animation-stopped animate-spin-base animation-delay-1 border-t-primary-600 border-x-transparent border-b-transparent"
       aria-hidden="true"
     />
     <div
-      class="absolute w-6 h-6 bg-transparent border-2 rounded-[50%] is-animation-stopped animate-spin-base animation-delay-2 border-t-primary-600 border-x-transparent border-b-transparent"
+      :class="{
+        'w-4 h-4 border-2': size === 'sm',
+        'w-6 h-6 border-2': size === 'base',
+        'w-10 h-10 border-2': size === 'lg',
+        'w-14 h-14 border-[3px]': size === 'xl',
+        'w-24 h-24 border-4': size === '2xl',
+        'w-48 h-48 border-8': size === '3xl',
+      }"
+      class="absolute bg-transparent rounded-[50%] is-animation-stopped animate-spin-base animation-delay-2 border-t-primary-600 border-x-transparent border-b-transparent"
       aria-hidden="true"
     />
     <div
-      class="absolute w-6 h-6 bg-transparent border-2 rounded-[50%] is-animation-stopped animate-spin-base animation-delay-3 border-t-primary-600 border-x-transparent border-b-transparent"
+      :class="{
+        'w-4 h-4': size === 'sm',
+        'w-6 h-6 border-2': size === 'base',
+        'w-10 h-10 border-2': size === 'lg',
+        'w-14 h-14 border-[3px]': size === 'xl',
+        'w-24 h-24 border-4': size === '2xl',
+        'w-48 h-48 border-8': size === '3xl',
+      }"
+      class="absolute bg-transparent rounded-[50%] is-animation-stopped animate-spin-base animation-delay-3 border-t-primary-600 border-x-transparent border-b-transparent"
       aria-hidden="true"
     />
   </div>
@@ -34,6 +82,22 @@
 
 <script>
 export default {
-  name: 'SpinnerBase'
+  name: 'SpinnerBase',
+  props: {
+    size: {
+      type: String,
+      default: 'base',
+      validator(value) {
+        return [
+          'sm',
+          'base',
+          'lg',
+          'xl',
+          '2xl',
+          '3xl'
+        ].includes(value);
+      }
+    }
+  }
 };
 </script>